+++ src/main/java/org/sca/calontir/cmpe/dto/AuthType.java
@@ -10,7 +10,7 @@ public class AuthType implements Serializable {
     private Long authTypeId;
     private String code;
     private String description;
-	private Integer orderValue;
+	private Long orderValue;
 
     public Long getAuthTypeId() {
         return authTypeId;
@@ -36,11 +36,11 @@ public class AuthType implements Serializable {
         this.description = description;
     }
 
-	public Integer getOrderValue() {
+	public Long getOrderValue() {
 		return orderValue;
 	}
 
-	public void setOrderValue(Integer orderValue) {
+	public void setOrderValue(Long orderValue) {
 		this.orderValue = orderValue;
 	}
