+++ src/main/java/org/sca/calontir/cmpe/utils/MarshalUtils.java
@@ -140,8 +140,8 @@ public class MarshalUtils {
 
 		@Override
 		public int compare(Authorization l, Authorization r) {
-			int left = l.getOrderValue() == null ? 99 : l.getOrderValue();
-			int right = r.getOrderValue() == null ? 99 : r.getOrderValue();
+			long left = l.getOrderValue() == null ? 99 : l.getOrderValue();
+			long right = r.getOrderValue() == null ? 99 : r.getOrderValue();
 			return (left < right ? -1 : (left == right ? 0 : 1));
 		}
 	}
